chore: update ESLint configuration, package.json, and add TypeScript build setup
- Modified ESLint config to allow `tsup.config.ts` as a default project for type-aware rules.
- Bumped package version to 1.0.0 and changed build script to use `tsup`.
- Added `tsup` as a dependency and updated lockfile.
- Introduced new workspace configuration for pnpm.
- Created initial `tsup.config.ts` for TypeScript build setup.
- Refactored CLI code to support new command structure with separate command files for `doctor`, `init`, and `uninstall`.
- Added utility functions for system checks and settings management.
